refactor(json-infrastructure): update package for the JsonCodecHelper class
