+++ infrastructure/json/src/main/java/io/github/aivruu/teams/player/infrastructure/json/PlayerJsonInfrastructureAggregateRootRepository.java
@@ -16,9 +16,9 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 package io.github.aivruu.teams.player.infrastructure.json;
 
-import io.github.aivruu.teams.persistence.infrastructure.utils.JsonCodecHelper;
 import io.github.aivruu.teams.player.domain.PlayerAggregateRoot;
 import io.github.aivruu.teams.shared.infrastructure.InfrastructureAggregateRootRepository;
+import io.github.aivruu.teams.shared.infrastructure.util.JsonCodecHelper;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
